% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighter.R
\name{runExample1}
\alias{runExample1}
\title{Near-perfect fit, to two target distributions, found in a series of steps.}
\usage{
runExample1()
}
\description{
Near-perfect fit, to two target distributions, found in a series of steps.
}

