% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighter.R
\name{weight.data}
\alias{weight.data}
\title{Re-weight the data to fit a list of target distributions.}
\usage{
weight.data(data.frame, target.distributions,
  data.frame.weight.name = default.weight.name,
  target.distribution.weight.names = default.weight.name, epsilon = 0.01,
  max.steps = 10)
}
\arguments{
\item{data.frame}{(data.frame) the data.frame with source data.}

\item{target.distributions}{(list of data.frame) one more target distributions
that the re-weighted data should approximate.}

\item{data.frame.weight.name}{(character) the name of the weight variable
in the source \code{data.frame}.}

\item{target.distribution.weight.names}{(character) the names of weight variables
in the \code{target.distributions} data frames.}

\item{epsilon}{(numeric) the target precision of the approximation.}

\item{max.steps}{(numeric) the maximum number of times the algorithm will
re-weight the data to find the best fit.}
}
\description{
\code{weight.data} iteratively re-weights the data to find the best fit
    to a list of target distributions.
}

